Add files via upload
Restaurant Website  
A simple and responsive restaurant website with a modern design. It includes a homepage, customer reviews, and a gallery of food and ambiance images, perfect for promoting any restaurant online!
echnologies Used
HTML5: For structuring the content.
CSS3: For designing the layout and styles.
Media Assets: High-quality images to enrich the website.